#68 - Moved the modal styles
